trading_bot/requirements.txt
pandas
numpy
matplotlib
ta
ccxt
alpaca-trade-api
ib_insync
streamlit
PyYAML
scikit-learn
requests
websocket-client
cryptography
yfinance